+++ src/components/SignInScreen/SignInScreen.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { shallow } from "enzyme";
 import { useSelector, useDispatch } from "react-redux";
-import { SignInScreen } from "components/SigninScreen/SignInScreen";
+import { SignInScreen } from "components/SignInScreen/SignInScreen";
 import { UserState } from "../../services/user/user";
 import { BrowserRouter } from "react-router-dom";
